Make sure last_activity and last_visit are timestamps
Could trigger a SQL error in the forums with manipulated activity cookies

ref: https://hackerone.com/reports/337620

+++ system/ee/legacy/libraries/Session.php
@@ -506,7 +506,7 @@ class EE_Session {
 		$expire = (60*60*24*365);
 
 		// Has the user been active before? If not we set the "last_activity" to the current time.
-		$this->sdata['last_activity'] = ( ! ee()->input->cookie('last_activity')) ? ee()->localize->now : ee()->input->cookie('last_activity');
+		$this->sdata['last_activity'] = (int) ( ! ee()->input->cookie('last_activity')) ? ee()->localize->now : ee()->input->cookie('last_activity');
 
 		// Is the "last_visit" cookie set?  If not, we set the last visit
 		// date to ten years ago. This is a kind of funky thing to do but
@@ -520,7 +520,7 @@ class EE_Session {
 		}
 		else
 		{
-			$this->userdata['last_visit'] = ee()->input->cookie('last_visit');
+			$this->userdata['last_visit'] = (int) ee()->input->cookie('last_visit');
 		}
 
 		// If the user has been inactive longer than the session length we'll
